+++ src/riace_ivn/__init__.py
@@ -1,11 +1,16 @@
 """Riace Bronzes IVN soft evidence aggregation."""
 
 from .aggregate import aggregate_v7_results, aggregate_v9_synthesis
+from .experimental_refinements import IVNState, delphi_interval, entropy_weights, topsis_rank
 from .model import compute_link_state, quality_interval
 
 __all__ = [
+    "IVNState",
     "aggregate_v7_results",
     "aggregate_v9_synthesis",
     "compute_link_state",
+    "delphi_interval",
+    "entropy_weights",
     "quality_interval",
+    "topsis_rank",
 ]